frontend/requirements.txt
streamlit
streamlit-folium
folium
geopandas
requests
plotly